+++ bindings/python/project/binlex/clients.pyi
@@ -65,16 +65,11 @@ class Web:
         url: str | None = None,
         verify: bool | None = None,
         api_key: str | None = None,
-        token: str | None = None,
     ) -> None: ...
     def url(self) -> str: ...
     def verify(self) -> bool: ...
     def api_key(self) -> str | None: ...
     def set_api_key(self, api_key: str | None) -> None: ...
-    def token(self) -> str | None: ...
-    def set_token(self, token: str | None) -> None: ...
-    def create_token(self) -> tuple[str, str]: ...
-    def clear_token(self, token: str) -> bool: ...
     def index_graph(
         self,
         sha256: str,
@@ -102,10 +97,6 @@ class Web:
     ) -> bool: ...
     def commit_index(self) -> bool: ...
     def clear_index(self) -> bool: ...
-    def sample_tags(self, sha256: str) -> list[str]: ...
-    def add_sample_tag(self, sha256: str, tag: str) -> bool: ...
-    def remove_sample_tag(self, sha256: str, tag: str) -> bool: ...
-    def replace_sample_tags(self, sha256: str, tags: list[str]) -> bool: ...
     def collection_tags(
         self,
         sha256: str,
