+++ bindings/python/project/binlex/indexing/local.pyi
@@ -146,16 +146,6 @@ class LocalIndex:
     def corpus_add(self, sha256: str, corpus: str) -> None: ...
     def corpus_replace(self, sha256: str, corpus: str) -> None: ...
     def corpus_rename(self, old_name: str, new_name: str) -> None: ...
-    def sample_tag_add(self, sha256: str, tag: str) -> None: ...
-    def sample_tag_remove(self, sha256: str, tag: str) -> None: ...
-    def sample_tag_replace(self, sha256: str, tags: list[str]) -> None: ...
-    def sample_tag_search(
-        self,
-        query: str,
-        page: int = 1,
-        page_size: int = 50,
-    ) -> TagSearchPage: ...
-    def sample_tag_list(self, sha256: str) -> list[str]: ...
     def sample_comment_add(
         self,
         sha256: str,
