+++ crates/binlex_tools/binlex-web/src/assets/js/api/client.js
@@ -32,6 +32,24 @@ async function postJson(url, payload) {
   return data;
 }
 
+async function deleteJson(url, payload) {
+  const response = await fetch(url, {
+    method: "DELETE",
+    headers: {
+      ...(payload !== undefined ? { "Content-Type": "application/json" } : {}),
+      "X-Requested-With": "binlex-web",
+      "Accept": "application/json",
+    },
+    credentials: "same-origin",
+    ...(payload !== undefined ? { body: JSON.stringify(payload) } : {}),
+  });
+  const data = await response.json().catch(() => ({}));
+  if (!response.ok) {
+    throw new Error(data?.error || "Request failed");
+  }
+  return data;
+}
+
 async function fetchJsonWithCredentials(url, options = {}) {
   const response = await fetch(url, {
     credentials: "same-origin",
@@ -53,6 +71,20 @@ async function fetchJsonWithCredentials(url, options = {}) {
 async function postJsonWithCredentials(url, payload) {
   return fetchJsonWithCredentials(url, {
     method: "POST",
+    body: JSON.stringify(payload),
+  });
+}
+
+async function putJsonWithCredentials(url, payload) {
+  return fetchJsonWithCredentials(url, {
+    method: "PUT",
+    body: JSON.stringify(payload),
+  });
+}
+
+async function deleteJsonWithCredentials(url, payload) {
+  return fetchJsonWithCredentials(url, {
+    method: "DELETE",
     body: JSON.stringify(payload),
   });
 }
