+++ crates/binlex_tools/binlex-web/src/includes/web_router.inc.rs
@@ -6,8 +6,6 @@ async fn run_server(
     let router = Router::new()
         .route("/", get(index_page))
         .route("/api/v1/version", get(version_api))
-        .route("/api/v1/token", post(create_token_api))
-        .route("/api/v1/token/clear", post(clear_token_api))
         .route("/api/v1/auth/bootstrap", post(auth_bootstrap_api))
         .route("/api/v1/auth/login", post(auth_login_api))
         .route("/api/v1/auth/login/2fa", post(auth_login_two_factor_api))
@@ -25,7 +23,7 @@ async fn run_server(
         .route("/api/v1/auth/password/reset", post(auth_password_reset_api))
         .route("/api/v1/auth/username/check", get(auth_username_check_api))
         .route("/api/v1/auth/me", get(auth_me_api))
-        .route("/api/v1/profile", get(profile_get_api))
+        .route("/api/v1/profile", get(profile_get_api).delete(profile_delete_api))
         .route("/api/v1/profile/password", post(profile_password_api))
         .route(
             "/api/v1/profile/2fa/setup",
@@ -55,19 +53,21 @@ async fn run_server(
             "/api/v1/profile/recovery/regenerate",
             post(profile_recovery_regenerate_api),
         )
-        .route("/api/v1/profile/delete", post(profile_delete_api))
         .route("/api/v1/admin/users", get(admin_users_api))
         .route("/api/v1/admin/users/create", post(admin_user_create_api))
         .route("/api/v1/admin/users/role", post(admin_user_role_api))
         .route("/api/v1/admin/users/enabled", post(admin_user_enabled_api))
         .route(
-            "/api/v1/admin/corpora/delete",
-            post(admin_delete_corpus_api),
+            "/api/v1/admin/corpora/{corpus}",
+            axum::routing::delete(admin_delete_corpus_api),
         )
-        .route("/api/v1/admin/tags/delete", post(admin_delete_tag_api))
         .route(
-            "/api/v1/admin/symbols/delete",
-            post(admin_delete_symbol_api),
+            "/api/v1/admin/tags/{tag}",
+            axum::routing::delete(admin_delete_tag_api),
+        )
+        .route(
+            "/api/v1/admin/symbols/{symbol}",
+            axum::routing::delete(admin_delete_symbol_api),
         )
         .route(
             "/api/v1/admin/users/password/reset",
@@ -78,8 +78,8 @@ async fn run_server(
             post(admin_user_key_regenerate_api),
         )
         .route(
-            "/api/v1/admin/users/picture/delete",
-            post(admin_user_picture_delete_api),
+            "/api/v1/admin/users/{username}/picture",
+            axum::routing::delete(admin_user_picture_delete_api),
         )
         .route(
             "/api/v1/admin/users/2fa/require",
@@ -93,7 +93,10 @@ async fn run_server(
             "/api/v1/admin/users/2fa/reset",
             post(admin_user_two_factor_reset_api),
         )
-        .route("/api/v1/admin/users/delete", post(admin_user_delete_api))
+        .route(
+            "/api/v1/admin/users/{username}",
+            axum::routing::delete(admin_user_delete_api),
+        )
         .route("/api/v1/index/graph", post(stage_index_graph))
         .route("/api/v1/index/function", post(stage_index_function))
         .route("/api/v1/index/block", post(stage_index_block))
@@ -101,41 +104,38 @@ async fn run_server(
         .route("/api/v1/index/commit", post(commit_index))
         .route("/api/v1/index/clear", post(clear_index))
         .route(
+            "/api/v1/index/sample",
+            post(upload).layer(DefaultBodyLimit::max(upload_limit)),
+        )
+        .route("/api/v1/index/status", get(upload_status))
+        .route(
             "/api/v1/corpora/collection",
-            get(get_collection_corpora_api),
+            get(get_collection_corpora_api)
+                .post(add_collection_corpus_api)
+                .delete(remove_collection_corpus_api),
         )
-        .route(
-            "/api/v1/corpora/collection/add",
-            post(add_collection_corpus_api),
-        )
-        .route(
-            "/api/v1/corpora/collection/remove",
-            post(remove_collection_corpus_api),
-        )
-        .route("/api/v1/tags/sample", get(get_sample_tags))
-        .route("/api/v1/tags/sample/add", post(add_sample_tag))
-        .route("/api/v1/tags/sample/remove", post(remove_sample_tag))
-        .route("/api/v1/tags/sample/replace", post(replace_sample_tags))
         .route("/api/v1/tags/search", get(search_tags_api))
-        .route("/api/v1/tags/add", post(add_tag_api))
-        .route("/api/v1/tags/collection", get(get_collection_tags))
-        .route("/api/v1/tags/collection/add", post(add_collection_tag_api))
+        .route("/api/v1/tags", post(add_tag_api))
         .route(
-            "/api/v1/tags/collection/remove",
-            post(remove_collection_tag_api),
+            "/api/v1/tags/collection",
+            get(get_collection_tags)
+                .post(add_collection_tag_api)
+                .delete(remove_collection_tag_api)
+                .put(replace_collection_tags_api),
         )
-        .route(
-            "/api/v1/tags/collection/replace",
-            post(replace_collection_tags_api),
-        )
-        .route("/api/v1/tags/search/sample", get(search_sample_tags_api))
         .route(
             "/api/v1/tags/search/collection",
             get(search_collection_tags_api),
         )
-        .route("/api/v1/symbols/collection", get(get_collection_symbols))
+        .route(
+            "/api/v1/symbols/collection",
+            get(get_collection_symbols)
+                .post(add_collection_symbol_api)
+                .delete(remove_collection_symbol_api)
+                .put(replace_collection_symbols_api),
+        )
         .route("/api/v1/symbols/search", get(search_symbols_api))
-        .route("/api/v1/symbols/add", post(add_symbol_api))
+        .route("/api/v1/symbols", post(add_symbol_api))
         .route("/api/v1/comments", get(get_entity_comments_api))
         .route("/api/v1/comments/add", post(add_entity_comment_api))
         .route(
@@ -143,28 +143,10 @@ async fn run_server(
             axum::routing::delete(delete_entity_comment_api),
         )
         .route("/api/v1/admin/comments", get(admin_comments_api))
-        .route(
-            "/api/v1/symbols/collection/add",
-            post(add_collection_symbol_api),
-        )
-        .route(
-            "/api/v1/symbols/collection/remove",
-            post(remove_collection_symbol_api),
-        )
-        .route(
-            "/api/v1/symbols/collection/replace",
-            post(replace_collection_symbols_api),
-        )
         .route("/api/v1/search", post(search_api))
         .route("/api/v1/search/detail", get(search_detail_api))
-        .route("/api/v1/action/yara", post(action_yara_api))
-        .route("/api/v1/corpora/add", post(add_corpus_api))
-        .route("/api/v1/corpora", get(search_corpora_api))
-        .route(
-            "/api/v1/upload/sample",
-            post(upload).layer(DefaultBodyLimit::max(upload_limit)),
-        )
-        .route("/api/v1/upload/status", get(upload_status))
+        .route("/api/v1/yara/render", post(action_yara_api))
+        .route("/api/v1/corpora", get(search_corpora_api).post(add_corpus_api))
         .route("/api/v1/download/sample", get(download_sample))
         .route("/api/v1/download/samples", get(download_samples))
         .route("/api/v1/download/json", get(download_json))
