+++ crates/binlex_tools/binlex-web/src/includes/web_state.inc.rs
@@ -8,6 +8,13 @@ struct AppState {
     staged_indexes: Arc<Mutex<BTreeMap<String, LocalIndex>>>,
 }
 
+#[derive(Clone, Copy, Debug, PartialEq, Eq)]
+enum RouteAccessPolicy {
+    Public,
+    Authenticated,
+    Admin,
+}
+
 impl AppState {
     fn two_factor_required(&self) -> bool {
         self.ui.auth.two_factor.enabled && self.ui.auth.two_factor.required
@@ -38,39 +45,49 @@ impl AppState {
             .ok_or_else(|| AppError::new("no staged indexing work for token"))
     }
 
-    fn route_auth_rule(&self, path: &str) -> Option<&WebAuthRuleConfig> {
-        self.ui.auth.rules.iter().find(|rule| {
-            rule.path == path || (rule.path.ends_with('/') && path.starts_with(&rule.path))
-        })
-    }
-
-    fn route_auth_enabled(&self, path: &str) -> bool {
+    fn route_access_policy(&self, method: &axum::http::Method, path: &str) -> RouteAccessPolicy {
         if !self.ui.auth.enabled {
-            return false;
+            return RouteAccessPolicy::Public;
         }
-        self.route_auth_rule(path)
-            .map(|rule| rule.enabled)
-            .unwrap_or(false)
-    }
-
-    fn route_auth_roles(&self, path: &str) -> Vec<String> {
-        self.route_auth_rule(path)
-            .map(|rule| rule.roles.clone())
-            .unwrap_or_default()
-    }
-
-    fn route_token_enabled(&self, path: &str) -> bool {
-        if !self.ui.token.enabled {
-            return false;
+        if path.starts_with("/api/v1/admin/")
+            || (method == axum::http::Method::POST && path == "/api/v1/corpora")
+            || (path.starts_with("/api/v1/comments/") && method == axum::http::Method::DELETE)
+        {
+            return RouteAccessPolicy::Admin;
         }
-        self.ui
-            .token
-            .rules
-            .iter()
-            .find(|rule| {
-                rule.path == path || (rule.path.ends_with('/') && path.starts_with(&rule.path))
-            })
-            .map(|rule| rule.enabled)
-            .unwrap_or(false)
+        if matches!(
+            (method, path),
+            (&axum::http::Method::POST, "/api/v1/index/graph")
+                | (&axum::http::Method::POST, "/api/v1/index/function")
+                | (&axum::http::Method::POST, "/api/v1/index/block")
+                | (&axum::http::Method::POST, "/api/v1/index/instruction")
+                | (&axum::http::Method::POST, "/api/v1/index/commit")
+                | (&axum::http::Method::POST, "/api/v1/index/clear")
+                | (&axum::http::Method::POST, "/api/v1/index/sample")
+                | (&axum::http::Method::POST, "/api/v1/corpora/collection")
+                | (&axum::http::Method::DELETE, "/api/v1/corpora/collection")
+                | (&axum::http::Method::POST, "/api/v1/tags")
+                | (&axum::http::Method::POST, "/api/v1/tags/collection")
+                | (&axum::http::Method::DELETE, "/api/v1/tags/collection")
+                | (&axum::http::Method::PUT, "/api/v1/tags/collection")
+                | (&axum::http::Method::POST, "/api/v1/symbols")
+                | (&axum::http::Method::POST, "/api/v1/symbols/collection")
+                | (&axum::http::Method::DELETE, "/api/v1/symbols/collection")
+                | (&axum::http::Method::PUT, "/api/v1/symbols/collection")
+                | (&axum::http::Method::POST, "/api/v1/comments/add")
+                | (&axum::http::Method::GET, "/api/v1/profile")
+                | (&axum::http::Method::DELETE, "/api/v1/profile")
+                | (&axum::http::Method::POST, "/api/v1/profile/password")
+                | (&axum::http::Method::POST, "/api/v1/profile/2fa/setup")
+                | (&axum::http::Method::POST, "/api/v1/profile/2fa/enable")
+                | (&axum::http::Method::POST, "/api/v1/profile/2fa/disable")
+                | (&axum::http::Method::POST, "/api/v1/profile/picture")
+                | (&axum::http::Method::DELETE, "/api/v1/profile/picture")
+                | (&axum::http::Method::POST, "/api/v1/profile/key/regenerate")
+                | (&axum::http::Method::POST, "/api/v1/profile/recovery/regenerate")
+        ) {
+            return RouteAccessPolicy::Authenticated;
+        }
+        RouteAccessPolicy::Public
     }
 }
