+++ crates/binlex_tools/binlex-web/templates/pages/partials/upload_corpus_picker.html
@@ -1,4 +1,4 @@
-<div class="upload-corpus-picker" data-upload-corpus-root="1" data-options='{{ options_json }}' data-selected='{{ selected_json }}'>
+<div class="upload-corpus-picker" data-upload-corpus-root="1" data-options='{{ options_json }}' data-selected='{{ selected_json }}' data-default-corpus="{{ default_corpus }}" data-locked="{% if locked %}true{% else %}false{% endif %}">
   <div class="upload-corpus-header">Corpora</div>
   <div class="upload-corpus-grid">
     <div class="upload-corpus-column">
