+++ src/databases/localdb/tags.rs
@@ -1,130 +1,6 @@
 use super::*;
 
 impl LocalDB {
-    pub fn sample_tag_add(&self, tag: &SampleTagRecord) -> Result<(), Error> {
-        let tag_name = normalize_metadata_name("tag", &tag.tag)?;
-        self.sqlite.execute(
-            "INSERT INTO sample_tags (sha256, tag, timestamp)
-             VALUES (?1, ?2, ?3)
-             ON CONFLICT(sha256, tag) DO UPDATE SET
-               timestamp = excluded.timestamp",
-            &[
-                SQLiteValue::Text(tag.sha256.clone()),
-                SQLiteValue::Text(tag_name),
-                SQLiteValue::Text(tag.timestamp.clone()),
-            ],
-        )?;
-        Ok(())
-    }
-
-    pub fn sample_tag_remove(&self, sha256: &str, tag: &str) -> Result<(), Error> {
-        self.sqlite.execute(
-            "DELETE FROM sample_tags WHERE sha256 = ?1 AND tag = ?2",
-            &[
-                SQLiteValue::Text(sha256.to_string()),
-                SQLiteValue::Text(tag.to_string()),
-            ],
-        )?;
-        Ok(())
-    }
-
-    pub fn sample_tag_list(&self, sha256: &str) -> Result<Vec<String>, Error> {
-        let rows = self.sqlite.query(
-            "SELECT tag
-             FROM sample_tags
-             WHERE sha256 = ?1
-             ORDER BY tag ASC",
-            &[SQLiteValue::Text(sha256.to_string())],
-        )?;
-        rows.into_iter()
-            .map(|row| {
-                row.get("tag")
-                    .and_then(|value| value.as_str())
-                    .ok_or_else(|| Error("sample tag row is missing tag".to_string()))
-                    .map(ToString::to_string)
-            })
-            .collect()
-    }
-
-    pub fn sample_tag_replace(
-        &self,
-        sha256: &str,
-        tags: &[String],
-        timestamp: &str,
-    ) -> Result<(), Error> {
-        self.sqlite.execute(
-            "DELETE FROM sample_tags WHERE sha256 = ?1",
-            &[SQLiteValue::Text(sha256.to_string())],
-        )?;
-        for tag in tags {
-            self.sample_tag_add(&SampleTagRecord {
-                sha256: sha256.to_string(),
-                tag: tag.clone(),
-                timestamp: timestamp.to_string(),
-            })?;
-        }
-        Ok(())
-    }
-
-    pub fn sample_tag_search(
-        &self,
-        query: &str,
-        page: usize,
-        page_size: usize,
-    ) -> Result<Page<SampleTagRecord>, Error> {
-        let page = page.max(1);
-        let page_size = page_size.max(1);
-        let offset = (page - 1) * page_size;
-        let limit = page_size + 1;
-        let pattern = if query.trim().is_empty() {
-            "%".to_string()
-        } else {
-            format!("%{}%", query.trim().to_ascii_lowercase())
-        };
-        let rows = self.sqlite.query(
-            "SELECT sha256, tag, timestamp
-             FROM sample_tags
-             WHERE LOWER(tag) LIKE ?1
-             ORDER BY tag ASC, sha256 ASC
-             LIMIT ?2 OFFSET ?3",
-            &[
-                SQLiteValue::Text(pattern),
-                SQLiteValue::Integer(limit as i64),
-                SQLiteValue::Integer(offset as i64),
-            ],
-        )?;
-        let has_next = rows.len() > page_size;
-        let items = rows
-            .into_iter()
-            .take(page_size)
-            .map(|row| -> Result<SampleTagRecord, Error> {
-                Ok(SampleTagRecord {
-                    sha256: row
-                        .get("sha256")
-                        .and_then(|value| value.as_str())
-                        .ok_or_else(|| Error("sample tag row is missing sha256".to_string()))?
-                        .to_string(),
-                    tag: row
-                        .get("tag")
-                        .and_then(|value| value.as_str())
-                        .ok_or_else(|| Error("sample tag row is missing tag".to_string()))?
-                        .to_string(),
-                    timestamp: row
-                        .get("timestamp")
-                        .and_then(|value| value.as_str())
-                        .ok_or_else(|| Error("sample tag row is missing timestamp".to_string()))?
-                        .to_string(),
-                })
-            })
-            .collect::<Result<Vec<_>, _>>()?;
-        Ok(Page {
-            items,
-            page,
-            page_size,
-            has_next,
-        })
-    }
-
     pub fn tag_add(
         &self,
         tag: &str,
@@ -244,10 +120,6 @@ impl LocalDB {
 
     pub fn tag_delete_global(&self, tag: &str) -> Result<bool, Error> {
         let tag = normalize_metadata_name("tag", tag)?;
-        self.sqlite.execute(
-            "DELETE FROM sample_tags WHERE tag = ?1",
-            &[SQLiteValue::Text(tag.clone())],
-        )?;
         self.sqlite.execute(
             "DELETE FROM collection_tags WHERE tag = ?1",
             &[SQLiteValue::Text(tag.clone())],
