+++ src/lib.rs
@@ -38,9 +38,9 @@ pub mod lifters;
 pub mod math;
 pub mod metadata;
 pub mod processor;
+pub mod query;
 pub mod rules;
 pub mod runtime;
-pub mod search;
 pub mod server;
 pub mod storage;
 pub mod util;
@@ -53,5 +53,5 @@ pub use core::Magic;
 pub use core::OperatingSystem;
 pub use core::Transport;
 pub use math::entropy;
-pub use search::Query;
+pub use query::Query;
 pub use util::hexdump;
