--- src/search/mod.rs
@@ -1,8 +1,0 @@
-pub mod query;
-
-pub use query::{
-    Query, QueryAnalysis, QueryCollection, QueryCompletionSpec, QueryError, QueryExpr, QueryField,
-    QueryTerm, SearchRoot, query_architecture_values, query_bool_matches, query_collection_values,
-    query_completion_specs, query_float_matches, query_integer_matches, query_score_matches,
-    query_size_matches, query_timestamp_matches,
-};
